third_party/blink/web_tests/external/wpt/service-workers/service-worker/navigate-window.https-expected.txt

This is a testharness.js-based test.
FAIL Clients.matchAll() should not show an old window as controlled after it navigates. assert_unreached: unexpected rejection: assert_equals: client should have expected frame type expected "auxiliary" but got "top-level" Reached unreachable code
FAIL Clients.matchAll() should not show an old window after it navigates. assert_unreached: unexpected rejection: assert_equals: client should have expected frame type expected "auxiliary" but got "top-level" Reached unreachable code
Harness: the test ran to completion.
